Create expection for unimplemented methods scene
